feat(create-dish): resolve problems and refactoring project structure
