+++ src/main/java/com/example/controller/command/formCommandsController.java
@@ -1,0 +1,4 @@
+package com.example.controller.command;
+
+public class formCommandsController {
+}
