+++ src/main/java/com/example/controller/command/listCommandsController.java
@@ -1,0 +1,4 @@
+package com.example.controller.command;
+
+public class listCommandsController {
+}
