+++ src/main/java/com/example/controller/dish/listDishController.java
@@ -1,0 +1,4 @@
+package com.example.controller.dish;
+
+public class listDishController {
+}
