+++ src/main/java/com/example/controller/home/HomeController.java
@@ -1,0 +1,4 @@
+package com.example.controller.home;
+
+public class HomeController {
+}
