+++ src/main/java/com/example/controller/table/listTablesController.java
@@ -1,0 +1,4 @@
+package com.example.controller.table;
+
+public class listTablesController {
+}
